% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_192_occurrences.R
\name{edit_192_occurrences}
\alias{edit_192_occurrences}
\title{Count 192 trinucleotide RNA edit occurrences}
\usage{
edit_192_occurrences(edit_context.mat, gr_sizes)
}
\arguments{
\item{gr_sizes}{A vector indicating the number of variants per GRanges}

\item{type_context}{result from type_context function}
}
\value{
Edit matrix with 192 trinucleotide RNA edit occurrences
}
\description{
@details
 This function is called by edit_matrix. It calculates the 192 trinucleotide context for all variants
 and then splits these per GRanges (samples). It then calculates how often each 192 trinucleotide context occurs.
 This function is a variation of mut_96_occcurences from MutationalPatterns
}
