% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRNAedits-data.R
\docType{data}
\name{edit_mat}
\alias{edit_mat}
\title{Example edit_matrix()}
\format{
A data frame of 192 rows and 2 columns that contains example RNA
edits
}
\description{
A data frame containing example RNA edits that can be used as input to
plot_192_edit_profile().
}
\details{
\itemize{ \item Sample - sample name \item chr - chromosome number \item pos -
chromosome position \item ref - reference base \item alt - alternate base }
}
\keyword{datasets}
