% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_type.R
\name{edit_type}
\alias{edit_type}
\title{Retrieve base RNA edit types from a VCF object
This function is a variation of mut_type from MutationalPatterns
A function to extract the base substitutions from a vcf and translate to
the 12 base substitution types. This function does not collapse G/A to C/U}
\usage{
edit_type(vcf)
}
\arguments{
\item{vcf}{A CollapsedVCF object}
}
\value{
Character vector with base substitution types
}
\description{
Retrieve base RNA edit types from a VCF object
This function is a variation of mut_type from MutationalPatterns
A function to extract the base substitutions from a vcf and translate to
the 12 base substitution types. This function does not collapse G/A to C/U
}
\examples{
## See the 'read_vcfs_as_granges()' example for how we obtained the
## following data:
vcfs <- readRDS(system.file("states/read_vcfs_as_granges_output.rds",
  package = "MutationalPatterns"
))

mut_type(vcfs[[1]])
}
\seealso{
\code{\link{read_vcfs_as_granges}}
}
