% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRNAedits-data.R
\docType{data}
\name{example.vcf.list}
\alias{example.vcf.list}
\title{RNA edit vcfs from two experiments.}
\format{
a granges obsejt with two vcf files
}
\description{
See the 'read_vcfs_as_granges()' example for how we obtained the
data.
}
\keyword{datasets}
