% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_192_edit_profile.R
\name{plot_192_edit_profile}
\alias{plot_192_edit_profile}
\title{Plot 192 trinucleotide RNA edit profile}
\usage{
plot_192_edit_profile(mut_matrix, colors = NA, ymax = 0.2, condensed = FALSE)
}
\arguments{
\item{mut_matrix}{192 trinucleotide profile matrix}

\item{colors}{Optional 12 value color vector.}

\item{ymax}{Y axis maximum value, default = 0.2}

\item{condensed}{More condensed plotting format. Default = F.}
}
\value{
192 trinucleotide profile plot
}
\description{
Plot relative contribution of 192 trinucleotides
This function is a variation of plot_96_proile from MutationalPatterns
}
\examples{
## See the 'edit_matrix()' example for how we obtained the
## RNA edit matrix information:
edit_mat <- load(system.file("data/edit_mat.rda"))

## Plot the 192-profile of two samples
plot_192_edit_profile(edit_mat, ymax=0.15)


}
